Rearranged functions and removed notification text at startup
Notification function currently invoked with the save button. I'm sure that there will be more menu items in the left menu pane on the main screen, but I just used that as a button to present the phone number dialog modally. In the future it would be a good idea to tie in the getNearby function with the getLandmarks function in ARViewController and its location data.